Refactor lower camel case python files to snake case
