+++ dash_bio/utils/ideogram_parser.py
