+++ tests/dashbio_demos/app_ideogram.py
@@ -3,7 +3,7 @@ import dash_core_components as dcc
 from dash.dependencies import Input, Output, State
 import dash_html_components as html
 import dash_bio
-from dash_bio.utils import ideogramParser as ideoParser
+from dash_bio.utils import ideogram_parser as ideoParser
 
 # running directly with Python
 if __name__ == '__main__':
