+++ tests/dashbio_demos/app_molecule3d.py
@@ -7,7 +7,7 @@ from dash.dependencies import Input, Output, State
 import dash_html_components as html
 import dash_core_components as dcc
 import dash_bio
-from dash_bio.utils import pdbParser as parser, stylesParser as sparser
+from dash_bio.utils import pdb_parser as parser, styles_parser as sparser
 
 # running directly with Python
 if __name__ == '__main__':
