+++ tests/dashbio_demos/app_needle_plot.py
@@ -8,9 +8,9 @@ import dash_html_components as html
 from dash.dependencies import Input, Output, State
 import dash_bio
 
-from dash_bio.utils.uniprotDatabaseTools import UniprotQueryBuilder
+from dash_bio.utils.uniprot_database_tools import UniprotQueryBuilder
 
-from dash_bio.utils.mutationDataParser import EMPTY_MUT_DATA, \
+from dash_bio.utils.mutation_data_parser import EMPTY_MUT_DATA, \
     load_protein_domains, parse_mutation_upload_file, \
     parse_domain_upload_file, parse_mutations_uniprot_data, load_mutation_data
