+++ tests/dashbio_demos/app_speck.py
@@ -3,7 +3,7 @@ from dash.dependencies import Input, Output, State
 import dash_html_components as html
 import dash_core_components as dcc
 import dash_bio
-from dash_bio.utils.xyzReader import read_xyz
+from dash_bio.utils.xyz_reader import read_xyz
 
 # running directly with Python
 if __name__ == '__main__':
